docker/base/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
numpy==1.17.2
Pillow==6.2.1
pyzmq==18.1.0
opencv-python==4.1.1.26
dataclasses==0.6
